Fix race conditions in enterprise originate
The multithreaded code for enterprise originate did not take into
account that reads and writes from one thread do not necessarily occur
in the same order as other threads. This can very occasionally result in
an enterprise originate call with a return status of
`SWITCH_STATUS_SUCCESS`, but with `NULL` for `*bleg`.

This patch changes the enterprise originate implementaiton so that (for
the most part) reads and writes to the shared `handle` happen while the
mutex locked, providing consistency between the threads.